+++ test.js
@@ -1,140 +1,158 @@
-const sinon = require('sinon')
-const assert = require('chai').assert
-const it = require('mocha').it
-const describe = require('mocha').describe
-const Wrapper = require('./')
+const {assert} = require('chai');
+const {describe, it, before} = require('mocha');
+const Wrapper = require('./');
 
-const connection = {
-  subscribe: function () {},
-  publish: function () {},
-  requestOne: function () {},
-  unsubscribe: function () {},
-  close: function () {},
-  currentServer: {url: {host: 'mock host'}}
-}
-const NATSrequestOne = sinon.stub(connection, 'requestOne')
-const NATSsubscribe = sinon.stub(connection, 'subscribe')
-const NATSpublish = sinon.stub(connection, 'publish')
-const NATSunsubscribe = sinon.stub(connection, 'unsubscribe')
-const NATSclose = sinon.stub(connection, 'close')
-const wrapper = Wrapper({connection: connection})
 
-describe('request', function () {
-  it('performs a request and returns successful result unwrapped', function (done) {
-    NATSrequestOne.callsArgWith(4, JSON.stringify([null, {id: 'ACC001'}]))
-    wrapper.request('account.get', {id: 'ACC001'}, function (error, account) {
-      assert.isNull(error)
-      assert.equal(account.id, 'ACC001')
-      done()
-    })
-  })
+const wrapper = new Wrapper({url: 'nats://localhost:4222', group: 'tests', requestTimeout: 100}); // a nats server should be running
+let nats;
+wrapper.on('error', () => {
+    process.exit(1);
+});
 
-  it('performs a request and returns error', function (done) {
-    NATSrequestOne.callsArgWith(4, JSON.stringify([{message: 'account service error'}]))
-    wrapper.request('account.get', {id: 'ACC001'}, function (error, account) {
-      assert.equal(error.message, 'account service error')
-      assert.isNotOk(account)
-      done()
-    })
-  })
+describe('wrapper', () => {
 
-  it('performs a request and returns an empty response', function (done) {
-    NATSrequestOne.callsArgWith(4, JSON.stringify([null, null]))
-    wrapper.request('account.get', {id: 'ACC001'}, function (error, account) {
-      assert.isNull(error)
-      assert.isNull(account)
-      done()
-    })
-  })
+    before((done) => {
+        wrapper.on('connect', () => {
+            nats = wrapper._nats;
+            done();
+        });
+    });
 
-  it('performs a request and returns error from error.detail', function (done) {
-    NATSrequestOne.callsArgWith(4, JSON.stringify([{detail: 'account service error'}]))
-    wrapper.request('account.get', {id: 'ACC001'}, function (error, response) {
-      assert.equal(error.message, 'account service error')
-      assert.isNotOk(response)
-      done()
-    })
-  })
+    describe('request', () => {
 
-  it('performs a request and returns timeout error', function (done) {
-    NATSrequestOne.callsArgWith(4, {code: 'REQ_TIMEOUT'})
-    wrapper.request('account.get', {id: 'ACC001'}, function (error, response) {
-      assert.equal(error.message, 'response timeout')
-      assert.isNotOk(response)
-      done()
-    })
-  })
-})
+        before((done) => {
+            wrapper.listen('request.ok', (message, respond) => {
+                respond(null, message);
+            });
+            wrapper.listen('request.password', (message, respond) => {
+                respond(null, message);
+            });
+            wrapper.listen('request.error', (message, respond) => {
+                respond(new Error('service error'));
+            });
+            wrapper.listen('request.empty', (message, respond) => {
+                respond(null, null);
+            });
+            wrapper.listen('request.error.detail', (message, respond) => {
+                respond({detail: 'service error'}, null);
+            });
+            done();
+        });
 
-describe('subscribe', function () {
-  it('performs a response-subscription to a subject', function (done) {
-    NATSsubscribe.callsArgWith(1, JSON.stringify({data: {one: 'two'}}), null, 'test.event.happened')
-    wrapper.subscribe('test.event.happened', function (message, replyTo, subject) {
-      assert.equal(message.data.one, 'two')
-      assert.isNull(replyTo)
-      assert.equal(subject, 'test.event.happened')
-      done()
-    })
-  })
-})
+        it('performs a request and returns successful result', (done) => {
+            wrapper.request('request.ok', {foo: 'bar'}, (error, response) => {
+                assert.isNull(error);
+                assert.equal(response.foo, 'bar');
+                done();
+            })
+        });
 
-describe('listen', function () {
-  it('listens to point-to-point requests', function (done) {
-    NATSsubscribe.callsArgWith(2, JSON.stringify({data: {one: 'two'}}), 'REPLYTO_INBOX', 'test.data.requested')
-    wrapper.listen('test.data.requested', function (message, respond) {
-      assert.equal(message.data.one, 'two')
-      respond(null, {result: true})
-      done()
-    })
-  })
-})
+        it('performs a request and screens password field in logs', (done) => {
+            wrapper.request('request.password', {password: 'should not see this'}, (error, response) => {
+                assert.isNull(error);
+                assert.equal(response.password, 'should not see this');
+                done()
+            })
+        });
 
-describe('process', function () {
-  it('performs a process-subscription to a subject', function (done) {
-    NATSsubscribe.callsArgWith(2, JSON.stringify({data: {one: 'two'}}), null, 'test.event.happened')
-    wrapper.process('test.event.happened', function (message, subject) {
-      assert.equal(message.data.one, 'two')
-      assert.equal(subject, 'test.event.happened')
-      done()
-    })
-  })
-})
+        it('performs a request and returns error', (done) => {
+            wrapper.request('request.error', {foo: 'bar'}, (error, response) => {
+                assert.equal(error.message, 'service error');
+                assert.isNotOk(response);
+                done()
+            })
+        });
 
-describe('publish', function () {
-  it('performs publishing to a subject', function (done) {
-    NATSpublish.callsArgWith(2, null)
-    wrapper.publish('test.event.happened', {message: 'yay!'})
-    assert.isTrue(NATSpublish.called)
-    done()
-  })
-})
+        it('performs a request and returns an empty response', (done) => {
+            wrapper.request('request.empty', {foo: 'bar'}, (error, response) => {
+                assert.isNull(error);
+                assert.isNull(response);
+                done()
+            })
+        });
 
-describe('respond', function () {
-  it('responds with an error', function (done) {
-    NATSpublish.callsArgWith(2, null)
-    wrapper._respond('test.request', new Error('something bad happened'))
-    assert.isTrue(NATSpublish.called)
-    done()
-  })
+        it('performs a request and returns error from error.detail', (done) => {
+            wrapper.request('request.error.detail', {foo: 'bar'}, (error, response) => {
+                assert.equal(error.message, 'service error');
+                assert.isNotOk(response);
+                done()
+            })
+        });
 
-  it('responds normally', function (done) {
-    NATSpublish.callsArgWith(2, null)
-    wrapper._respond('test.request', null, {response: 'ok'})
-    assert.isTrue(NATSpublish.called)
-    done()
-  })
-})
+        it('performs a request and returns timeout error', (done) => {
+            wrapper.request('request.timeout', {foo: 'bar'}, (error, response) => {
+                assert.equal(error.message, 'response timeout');
+                assert.isNotOk(response);
+                done()
+            })
+        }).timeout(110);
+    });
 
-describe('unsubscribe', function () {
-  it('unsubscribes from NATS subject', function () {
-    wrapper.unsubscribe(123)
-    assert.isTrue(NATSunsubscribe.withArgs(123).calledOnce)
-  })
-})
+    describe('subscribe', () => {
+        it('performs a response-subscription to a subject', (done) => {
+            wrapper.subscribe('broadcast', (message, replyTo, subject) => {
+                assert.equal(message.foo, 'bar');
+                assert.isNotOk(replyTo);
+                assert.equal(subject, 'broadcast');
+                done()
+            });
+            wrapper.publish('broadcast', {foo: 'bar'});
+        })
+    });
 
-describe('close', function () {
-  it('closes underlying connection with NATS', function () {
-    wrapper.close()
-    assert.isTrue(NATSclose.calledOnce)
-  })
-})
+    describe('listen', () => {
+        it('listens to requests', (done) => {
+            wrapper.listen('request.listen', (message, respond) => {
+                assert.equal(message.foo, 'bar');
+                respond(null, {bar: 'foo'});
+            });
+            wrapper.request('request.listen', {foo: 'bar'}, (error, response) => {
+                assert.isNull(error);
+                assert.equal(response.bar, 'foo');
+                done();
+            });
+        })
+    });
+
+    describe('process', () => {
+        it('process a queue message as group member', (done) => {
+            wrapper.process('process', (message, subject) => {
+                assert.equal(message.foo, 'bar');
+                assert.equal(subject, 'process');
+                done()
+            });
+            wrapper.publish('process', {foo: 'bar'});
+        })
+    });
+
+    describe('publish', () => {
+        it('publishes a message to a subject', (done) => {
+            wrapper.process('publish', (message) => {
+                assert.equal(message.message, 'yay!');
+                done();
+            });
+            wrapper.publish('publish', {message: 'yay!'});
+        });
+    });
+
+    describe('unsubscribe', () => {
+        it('unsubscribes from NATS subject', () => {
+            const sid = wrapper.subscribe('subscribe', ()=>{});
+            wrapper.unsubscribe(sid);
+            wrapper._nats.on('unsubscribe', (subId, subject) => {
+                assert.equal(subId, sid);
+                assert.equal(subject, 'unsubscribe');
+                done();
+            });
+        })
+    });
+
+    describe('close', () => {
+        it('closes underlying connection with NATS', (done) => {
+            wrapper.close();
+            assert.isTrue(wrapper._nats.closed);
+            done();
+        })
+    });
+
+});