+++ CHANGELOG.md
@@ -1,6 +1,30 @@
+[2.0.1] - 2017-09-27
+--------------------
+- [x] UPDATED TS bindings to match js implementation
+
+
+[2.0.0] - 2017-09-15
+--------------------
+- [x] ADDED ids to request/response
+- [x] ADDED `connected` event handling fix #13
+- [x] CHANGE sinon mocks to live NATS connection @ tests
+- [x] REMOVED js standard
+
+
+[1.6.0] - 2017-08-02
+--------------------
+- [x] ADDED password screening fix #12
+
+
+[1.5.4] - 2017-07-03
+--------------------
+- [x] ADDED more info @ error response loggin
+
+
 [1.5.3] - 2017-05-21
 --------------------
 - [x] ADDED `unsubscribe` method
+- [x] ADDED more info @ error response logging
 
 
 [1.5.2] - 2017-05-17